evals: add widgets for generating eval input
This change adds a UI element where you can input project, tech spec,
PRD, and story details to get back a downloadable dataframe that
conforms to the annotation tool input format.

It also includes a cell at the end that captures all the annotated data
with notes for helping with error analysis.